spaces and traits
