+++ crates/ferrmion-core/src/spaces.rs
@@ -1,0 +1,5 @@
+//! Definitions of mathematical spaces.
+
+pub trait Fermion {}
+
+pub trait Qubit {}
